initial docker-compose based mesh
Creates a central instance and an executor instance that connects to
it.

Current pain point is that you need to use `docker port` to get the
external ports to be able to connect to central's mesh API and the
executor's published port 3001-3005 for the apps deployed.

Furthermore, if using boot2docker or some other remote docker, you need
to use $DOCKER_HOST to determine the IP.